java_updater: interval is integer seconds only

+++ aoirint_mcping_server/java_updater/java_updater.py
@@ -23,7 +23,7 @@ logger = logging.Logger(name="java_updater")
 
 class JavaUpdaterConfig(BaseModel):
     database_url: str
-    interval: float
+    interval: int
     timeout: float
 
 
@@ -115,7 +115,7 @@ def main() -> None:
     )
     parser.add_argument(
         "--interval",
-        type=float,
+        type=int,
         default=os.environ.get("MCPING_JAVA_UPDATER_INTERVAL", "300"),
     )
     parser.add_argument(
@@ -147,7 +147,7 @@ def main() -> None:
     log_level: int = args.log_level
     log_file: str | None = args.log_file
     database_url: str = args.database_url
-    interval: float = args.interval
+    interval: int = args.interval
     timeout: float = args.timeout
     loop: bool = args.loop
